requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
ortools==9.8.3296
requests==2.31.0
httpx==0.25.2
numpy==1.24.3
python-jose[cryptography]==3.3.0
tenacity==8.2.3
